Create batch responsive image sizes
Create two batches of blog images to feature the food
the recipes make. One batch is for desktop veiwing and
follows the responsivess based on device size/resolution.
The second batch is for art direction puroses and
may not be used. This batch is for optimised veiwing on
mobile devices.